refactor(FAT1-00): changing reset password path
