+++ lib/presentation/reset_password/views/widgets/build_profile_reset_password_form.dart
@@ -1,7 +1,7 @@
 import 'package:flowery_tracking_app/core/constants/app_text.dart';
 import 'package:flowery_tracking_app/domain/entities/requests/profile_reset_password/profile_reset_password_entity.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_cubit.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_intent.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_cubit.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_intent.dart';
 import 'package:flowery_tracking_app/utils/common_widgets/custom_elevated_button.dart';
 import 'package:flowery_tracking_app/utils/common_widgets/custom_text_form_field.dart';
 import 'package:flowery_tracking_app/utils/validations.dart';
