+++ lib/presentation/reset_password/views_model/profile_reset_password_cubit.dart
@@ -3,8 +3,8 @@ import 'package:flowery_tracking_app/api/responses/profile_reset_password/profil
 import 'package:flowery_tracking_app/core/state_status/state_status.dart';
 import 'package:flowery_tracking_app/domain/entities/requests/profile_reset_password/profile_reset_password_entity.dart';
 import 'package:flowery_tracking_app/domain/use_cases/profile_reset_password/profile_reset_password_usecase.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_intent.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_state.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_intent.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
