+++ test/presentation/reset_password/views/widgets/profile_reset_password_app_bar_test.dart
@@ -1,6 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowery_tracking_app/core/constants/app_text.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views/widgets/profile_reset_password_app_bar.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views/widgets/profile_reset_password_app_bar.dart';
 import 'package:flowery_tracking_app/utils/common_widgets/custom_app_bar.dart';
 import 'package:flowery_tracking_app/utils/common_widgets/custom_back_arrow.dart';
 import 'package:flutter/material.dart';
