+++ test/presentation/reset_password/views/widgets/profile_reset_password_body_test.dart
@@ -1,9 +1,9 @@
 import 'package:flowery_tracking_app/core/di/di.dart';
 import 'package:flowery_tracking_app/core/state_status/state_status.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views/widgets/build_profile_reset_password_form.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views/widgets/profile_reset_password_body.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_cubit.dart';
-import 'package:flowery_tracking_app/presentation/profile/reset_password/views_model/profile_reset_password_state.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views/widgets/build_profile_reset_password_form.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views/widgets/profile_reset_password_body.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_cubit.dart';
+import 'package:flowery_tracking_app/presentation/reset_password/views_model/profile_reset_password_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
